Ignore properties without public setters
