+++ src/Components/Endpoints/src/FormMapping/Metadata/FormDataTypeKind.cs
@@ -1,0 +1,12 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Microsoft.AspNetCore.Components.Endpoints.FormMapping.Metadata;
+
+internal enum FormDataTypeKind
+{
+    Primitive,
+    Collection,
+    Dictionary,
+    Object,
+}
