+++ src/Components/Endpoints/test/Binding/FormDataMetadataFactoryTests.cs
@@ -16,7 +16,7 @@ public class FormDataMetadataFactoryTests
         // Assert
         Assert.NotNull(metadata);
         Assert.Equal(typeof(Customer), metadata.Type);
-        Assert.Equal(TypeKind.Object, metadata.Kind);
+        Assert.Equal(FormDataTypeKind.Object, metadata.Kind);
         Assert.NotNull(metadata.Constructor);
         Assert.Collection(metadata.Properties,
             property =>
@@ -24,7 +24,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Id", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(int), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Empty(property.PropertyMetadata.Properties);
             },
@@ -33,7 +33,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Name", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(string), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Empty(property.PropertyMetadata.Properties);
             });
@@ -50,7 +50,7 @@ public class FormDataMetadataFactoryTests
         // Assert
         Assert.NotNull(metadata);
         Assert.Equal(typeof(CustomerWithAddress), metadata.Type);
-        Assert.Equal(TypeKind.Object, metadata.Kind);
+        Assert.Equal(FormDataTypeKind.Object, metadata.Kind);
         Assert.False(metadata.IsRecursive);
         Assert.NotNull(metadata.Constructor);
         Assert.Collection(metadata.Properties,
@@ -59,7 +59,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("BillingAddress", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(Address), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Object, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Object, property.PropertyMetadata.Kind);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.False(property.PropertyMetadata.IsRecursive);
                 Assert.Collection(property.PropertyMetadata.Properties,
@@ -68,7 +68,7 @@ public class FormDataMetadataFactoryTests
                        Assert.Equal("Street", property.Property.Name);
                        Assert.NotNull(property.PropertyMetadata);
                        Assert.Equal(typeof(string), property.PropertyMetadata.Type);
-                       Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                       Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                        Assert.Null(property.PropertyMetadata.Constructor);
                        Assert.Empty(property.PropertyMetadata.Properties);
                    },
@@ -77,7 +77,7 @@ public class FormDataMetadataFactoryTests
                        Assert.Equal("City", property.Property.Name);
                        Assert.NotNull(property.PropertyMetadata);
                        Assert.Equal(typeof(string), property.PropertyMetadata.Type);
-                       Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                       Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                        Assert.Null(property.PropertyMetadata.Constructor);
                        Assert.Empty(property.PropertyMetadata.Properties);
                    });
@@ -87,7 +87,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("ShippingAddress", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(Address), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Object, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Object, property.PropertyMetadata.Kind);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.False(property.PropertyMetadata.IsRecursive);
                 Assert.Collection(property.PropertyMetadata.Properties,
@@ -96,7 +96,7 @@ public class FormDataMetadataFactoryTests
                         Assert.Equal("Street", property.Property.Name);
                         Assert.NotNull(property.PropertyMetadata);
                         Assert.Equal(typeof(string), property.PropertyMetadata.Type);
-                        Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                        Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                         Assert.Null(property.PropertyMetadata.Constructor);
                         Assert.Empty(property.PropertyMetadata.Properties);
                     },
@@ -105,7 +105,7 @@ public class FormDataMetadataFactoryTests
                         Assert.Equal("City", property.Property.Name);
                         Assert.NotNull(property.PropertyMetadata);
                         Assert.Equal(typeof(string), property.PropertyMetadata.Type);
-                        Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                        Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                         Assert.Null(property.PropertyMetadata.Constructor);
                         Assert.Empty(property.PropertyMetadata.Properties);
                     });
@@ -123,7 +123,7 @@ public class FormDataMetadataFactoryTests
         // Assert
         Assert.NotNull(metadata);
         Assert.Equal(typeof(Address), metadata.Type);
-        Assert.Equal(TypeKind.Object, metadata.Kind);
+        Assert.Equal(FormDataTypeKind.Object, metadata.Kind);
         Assert.Null(metadata.Constructor);
         Assert.Collection(metadata.Properties,
             property =>
@@ -131,7 +131,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Street", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(string), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Empty(property.PropertyMetadata.Properties);
             },
@@ -140,7 +140,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("City", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(string), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Empty(property.PropertyMetadata.Properties);
             });
@@ -157,7 +157,7 @@ public class FormDataMetadataFactoryTests
         // Assert
         Assert.NotNull(metadata);
         Assert.Equal(typeof(KeyValuePair<int, string>), metadata.Type);
-        Assert.Equal(TypeKind.Object, metadata.Kind);
+        Assert.Equal(FormDataTypeKind.Object, metadata.Kind);
         Assert.NotNull(metadata.Constructor);
         Assert.Collection(metadata.ConstructorParameters,
             parameter =>
@@ -165,7 +165,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("key", parameter.Name);
                 Assert.NotNull(parameter.ParameterMetadata);
                 Assert.Equal(typeof(int), parameter.ParameterMetadata.Type);
-                Assert.Equal(TypeKind.Primitive, parameter.ParameterMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Primitive, parameter.ParameterMetadata.Kind);
                 Assert.Null(parameter.ParameterMetadata.Constructor);
                 Assert.Empty(parameter.ParameterMetadata.Properties);
             },
@@ -174,7 +174,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("value", parameter.Name);
                 Assert.NotNull(parameter.ParameterMetadata);
                 Assert.Equal(typeof(string), parameter.ParameterMetadata.Type);
-                Assert.Equal(TypeKind.Primitive, parameter.ParameterMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Primitive, parameter.ParameterMetadata.Kind);
                 Assert.Null(parameter.ParameterMetadata.Constructor);
                 Assert.Empty(parameter.ParameterMetadata.Properties);
             });
@@ -191,7 +191,7 @@ public class FormDataMetadataFactoryTests
 
         // Assert
         Assert.Equal(typeof(CustomerWithOrders), metadata.Type);
-        Assert.Equal(TypeKind.Object, metadata.Kind);
+        Assert.Equal(FormDataTypeKind.Object, metadata.Kind);
         Assert.NotNull(metadata.Constructor);
         Assert.Collection(metadata.Properties,
             property =>
@@ -199,7 +199,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Id", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(int), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Empty(property.PropertyMetadata.Properties);
             },
@@ -208,7 +208,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Name", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(string), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Empty(property.PropertyMetadata.Properties);
             },
@@ -217,7 +217,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Orders", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(List<CustomerWithOrders.Order>), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Collection, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Collection, property.PropertyMetadata.Kind);
                 Assert.NotNull(property.PropertyMetadata.ElementType);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Collection(
@@ -227,7 +227,7 @@ public class FormDataMetadataFactoryTests
                         Assert.Equal("Id", subProperty.Property.Name);
                         Assert.NotNull(subProperty.PropertyMetadata);
                         Assert.Equal(typeof(int), subProperty.PropertyMetadata.Type);
-                        Assert.Equal(TypeKind.Primitive, subProperty.PropertyMetadata.Kind);
+                        Assert.Equal(FormDataTypeKind.Primitive, subProperty.PropertyMetadata.Kind);
                         Assert.Null(subProperty.PropertyMetadata.Constructor);
                         Assert.Empty(subProperty.PropertyMetadata.Properties);
                     },
@@ -236,7 +236,7 @@ public class FormDataMetadataFactoryTests
                         Assert.Equal("ProductName", subProperty.Property.Name);
                         Assert.NotNull(subProperty.PropertyMetadata);
                         Assert.Equal(typeof(string), subProperty.PropertyMetadata.Type);
-                        Assert.Equal(TypeKind.Primitive, subProperty.PropertyMetadata.Kind);
+                        Assert.Equal(FormDataTypeKind.Primitive, subProperty.PropertyMetadata.Kind);
                         Assert.Null(subProperty.PropertyMetadata.Constructor);
                         Assert.Empty(subProperty.PropertyMetadata.Properties);
                     });
@@ -254,7 +254,7 @@ public class FormDataMetadataFactoryTests
         // Assert
         Assert.NotNull(metadata);
         Assert.Equal(typeof(CompanyWithWarehousesByLocation), metadata.Type);
-        Assert.Equal(TypeKind.Object, metadata.Kind);
+        Assert.Equal(FormDataTypeKind.Object, metadata.Kind);
         Assert.NotNull(metadata.Constructor);
         Assert.Collection(metadata.Properties,
             property =>
@@ -262,7 +262,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Name", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(string), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Empty(property.PropertyMetadata.Properties);
             },
@@ -271,7 +271,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Warehouses", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(Dictionary<string, Warehouse>), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Dictionary, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Dictionary, property.PropertyMetadata.Kind);
                 Assert.NotNull(property.PropertyMetadata.KeyType);
                 Assert.NotNull(property.PropertyMetadata.ValueType);
                 Assert.Null(property.PropertyMetadata.Constructor);
@@ -282,7 +282,7 @@ public class FormDataMetadataFactoryTests
                         Assert.Equal("Name", subProperty.Property.Name);
                         Assert.NotNull(subProperty.PropertyMetadata);
                         Assert.Equal(typeof(string), subProperty.PropertyMetadata.Type);
-                        Assert.Equal(TypeKind.Primitive, subProperty.PropertyMetadata.Kind);
+                        Assert.Equal(FormDataTypeKind.Primitive, subProperty.PropertyMetadata.Kind);
                         Assert.Null(subProperty.PropertyMetadata.Constructor);
                         Assert.Empty(subProperty.PropertyMetadata.Properties);
                     },
@@ -291,7 +291,7 @@ public class FormDataMetadataFactoryTests
                         Assert.Equal("Address", subProperty.Property.Name);
                         Assert.NotNull(subProperty.PropertyMetadata);
                         Assert.Equal(typeof(Address), subProperty.PropertyMetadata.Type);
-                        Assert.Equal(TypeKind.Object, subProperty.PropertyMetadata.Kind);
+                        Assert.Equal(FormDataTypeKind.Object, subProperty.PropertyMetadata.Kind);
                         Assert.Null(subProperty.PropertyMetadata.Constructor);
                         Assert.Collection(
                             subProperty.PropertyMetadata.Properties,
@@ -300,7 +300,7 @@ public class FormDataMetadataFactoryTests
                                 Assert.Equal("Street", subSubProperty.Property.Name);
                                 Assert.NotNull(subSubProperty.PropertyMetadata);
                                 Assert.Equal(typeof(string), subSubProperty.PropertyMetadata.Type);
-                                Assert.Equal(TypeKind.Primitive, subSubProperty.PropertyMetadata.Kind);
+                                Assert.Equal(FormDataTypeKind.Primitive, subSubProperty.PropertyMetadata.Kind);
                                 Assert.Null(subSubProperty.PropertyMetadata.Constructor);
                                 Assert.Empty(subSubProperty.PropertyMetadata.Properties);
                             },
@@ -309,7 +309,7 @@ public class FormDataMetadataFactoryTests
                                 Assert.Equal("City", subSubProperty.Property.Name);
                                 Assert.NotNull(subSubProperty.PropertyMetadata);
                                 Assert.Equal(typeof(string), subSubProperty.PropertyMetadata.Type);
-                                Assert.Equal(TypeKind.Primitive, subSubProperty.PropertyMetadata.Kind);
+                                Assert.Equal(FormDataTypeKind.Primitive, subSubProperty.PropertyMetadata.Kind);
                                 Assert.Null(subSubProperty.PropertyMetadata.Constructor);
                                 Assert.Empty(subSubProperty.PropertyMetadata.Properties);
                             });
@@ -328,7 +328,7 @@ public class FormDataMetadataFactoryTests
         // Assert
         Assert.NotNull(metadata);
         Assert.Equal(typeof(RecursiveList<string>), metadata.Type);
-        Assert.Equal(TypeKind.Object, metadata.Kind);
+        Assert.Equal(FormDataTypeKind.Object, metadata.Kind);
         Assert.True(metadata.IsRecursive);
         Assert.NotNull(metadata.Constructor);
         Assert.Collection(metadata.Properties,
@@ -337,7 +337,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Head", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(string), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Empty(property.PropertyMetadata.Properties);
             },
@@ -346,7 +346,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Tail", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(RecursiveList<string>), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Object, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Object, property.PropertyMetadata.Kind);
                 Assert.Same(metadata, property.PropertyMetadata);
             });
     }
@@ -363,7 +363,7 @@ public class FormDataMetadataFactoryTests
         Assert.NotNull(metadata);
         Assert.Equal(typeof(BaseList<string>), metadata.Type);
         Assert.True(metadata.IsRecursive);
-        Assert.Equal(TypeKind.Object, metadata.Kind);
+        Assert.Equal(FormDataTypeKind.Object, metadata.Kind);
         Assert.NotNull(metadata.Constructor);
         Assert.Collection(metadata.Properties,
             property =>
@@ -371,7 +371,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Head", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(string), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Empty(property.PropertyMetadata.Properties);
             },
@@ -381,7 +381,7 @@ public class FormDataMetadataFactoryTests
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(DerivedList<string>), property.PropertyMetadata.Type);
                 Assert.True(property.PropertyMetadata.IsRecursive);
-                Assert.Equal(TypeKind.Object, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Object, property.PropertyMetadata.Kind);
                 Assert.NotSame(metadata, property.PropertyMetadata);
                 Assert.Collection(property.PropertyMetadata.Properties,
                     subProperty =>
@@ -389,7 +389,7 @@ public class FormDataMetadataFactoryTests
                         Assert.Equal("Head", subProperty.Property.Name);
                         Assert.NotNull(subProperty.PropertyMetadata);
                         Assert.Equal(typeof(string), subProperty.PropertyMetadata.Type);
-                        Assert.Equal(TypeKind.Primitive, subProperty.PropertyMetadata.Kind);
+                        Assert.Equal(FormDataTypeKind.Primitive, subProperty.PropertyMetadata.Kind);
                         Assert.Null(subProperty.PropertyMetadata.Constructor);
                         Assert.Empty(subProperty.PropertyMetadata.Properties);
                     },
@@ -398,7 +398,7 @@ public class FormDataMetadataFactoryTests
                         Assert.Equal("Tail", subProperty.Property.Name);
                         Assert.NotNull(subProperty.PropertyMetadata);
                         Assert.Equal(typeof(DerivedList<string>), subProperty.PropertyMetadata.Type);
-                        Assert.Equal(TypeKind.Object, subProperty.PropertyMetadata.Kind);
+                        Assert.Equal(FormDataTypeKind.Object, subProperty.PropertyMetadata.Kind);
                         Assert.Same(property.PropertyMetadata, subProperty.PropertyMetadata);
                     });
             });
@@ -415,7 +415,7 @@ public class FormDataMetadataFactoryTests
         // Assert
         Assert.NotNull(metadata);
         Assert.Equal(typeof(Tree<string>), metadata.Type);
-        Assert.Equal(TypeKind.Object, metadata.Kind);
+        Assert.Equal(FormDataTypeKind.Object, metadata.Kind);
         Assert.True(metadata.IsRecursive);
         Assert.NotNull(metadata.Constructor);
         Assert.Collection(metadata.Properties,
@@ -424,7 +424,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Value", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(string), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Empty(property.PropertyMetadata.Properties);
             },
@@ -433,7 +433,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Children", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(List<Tree<string>>), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Collection, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Collection, property.PropertyMetadata.Kind);
                 Assert.False(property.PropertyMetadata.IsRecursive);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Empty(property.PropertyMetadata.Properties);
@@ -452,7 +452,7 @@ public class FormDataMetadataFactoryTests
         // Assert
         Assert.NotNull(metadata);
         Assert.Equal(typeof(DictionaryTree<string>), metadata.Type);
-        Assert.Equal(TypeKind.Object, metadata.Kind);
+        Assert.Equal(FormDataTypeKind.Object, metadata.Kind);
         Assert.True(metadata.IsRecursive);
         Assert.NotNull(metadata.Constructor);
         Assert.Collection(metadata.Properties,
@@ -461,7 +461,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Value", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(string), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Primitive, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Primitive, property.PropertyMetadata.Kind);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Empty(property.PropertyMetadata.Properties);
             },
@@ -470,7 +470,7 @@ public class FormDataMetadataFactoryTests
                 Assert.Equal("Children", property.Property.Name);
                 Assert.NotNull(property.PropertyMetadata);
                 Assert.Equal(typeof(Dictionary<string, DictionaryTree<string>>), property.PropertyMetadata.Type);
-                Assert.Equal(TypeKind.Dictionary, property.PropertyMetadata.Kind);
+                Assert.Equal(FormDataTypeKind.Dictionary, property.PropertyMetadata.Kind);
                 Assert.False(property.PropertyMetadata.IsRecursive);
                 Assert.Null(property.PropertyMetadata.Constructor);
                 Assert.Empty(property.PropertyMetadata.Properties);
